LayoutTests/http/tests/security/contentSecurityPolicy/inline-script-blocked-goofy.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="X-WebKit-CSP" content="script-src http://127.0.0.1:*; options goofy">
<script src="resources/dump-as-text.js"></script>
</head>
<body onload="alert('FAIL 3 of 3')">
This test passes if it doesn't alert fail.
<script>
alert('FAIL 1 of 3');
</script>
<iframe src="javascript:alert('FAIL 2 of 3')"></iframe>
</body>
</html>
